+++ src/client/session.rs
@@ -1,10 +1,13 @@
 //! SessionData
 
-use crate::HttpClient;
+use crate::{
+    error::{Result, PandaError},
+    models::{ gateway::commands::Command, user::StatusUpdate },
+    HttpClient};
 
 use std::sync::atomic::{AtomicBool, Ordering};
 
-use futures::lock::Mutex;
+use futures::{ channel::mpsc::UnboundedSender, lock::Mutex, sink::SinkExt };
 
 /// The struct of the current session of the bot.
 pub struct SessionData<S> {
@@ -12,15 +15,17 @@ pub struct SessionData<S> {
     pub http: HttpClient,
     pub state: S,
     is_resumable: AtomicBool,
+    to_gateway_ch: Mutex<UnboundedSender<Command>>
 }
 
 impl<S> SessionData<S> {
-    pub(crate) fn new(token: String, state: S) -> Self {
+    pub(crate) fn new(token: String, state: S, to_gateway_ch: UnboundedSender<Command>) -> Self {
         SessionData {
             id: Mutex::new("".into()),
             http: HttpClient::new(token),
             state,
             is_resumable: AtomicBool::new(true),
+            to_gateway_ch: Mutex::new(to_gateway_ch)
         }
     }
 
@@ -44,5 +49,16 @@ impl<S> SessionData<S> {
     pub(crate) async fn id(&self) -> String {
         let session_id = self.id.lock().await;
         session_id.clone()
+    }
+
+    //Send
+    pub async fn update_status(&self, status_update: StatusUpdate) -> Result<()> {
+
+        let cmd = Command::new_status_update(status_update);
+
+        // TODO: Into<Error>
+        self.to_gateway_ch.lock().await.send(cmd).await.map_err(|_| PandaError::ConnectionClosed)?;
+
+        Ok(())
     }
 }