+++ src/models/gateway/payload.rs
@@ -32,6 +32,7 @@ pub(crate) enum Opcode {
     Dispatch = 0,
     Heartbeat = 1,
     Identify = 2,
+    StatusUpdate = 3,
     Resume = 6,
     Reconnect = 7,
     RequestGuildMember = 8,