+++ src/models/user/mod.rs
@@ -1,7 +1,10 @@
 //! User related models
 
 mod activity;
+mod status_update;
+
 pub use activity::Activity;
+pub use status_update::StatusUpdate;
 
 use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
@@ -10,7 +13,8 @@ pub struct User {
     pub username: String,
     pub discriminator: String,
     pub avatar: Option<String>,
-    pub bot: Option<bool>,
+    #[serde(default)]
+    pub bot: bool,
     pub locale: Option<String>,
     pub verified: Option<bool>,
     pub email: Option<String>,