+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.5.3] ??????????????????
+## [0.5.3] - 2020-06-13
 ### Added
 - `Message.add_reaction()`, alias for `http.add_message_reaction()`
-
+- `Message.pin()`, alias for `http.add_pinned_channel_message()`
+- `Message.unpin()`, alias for `http.add_message_reaction()`
+- Now you can update the bot status
 
 ### Fixes
-- PresenceUpdate event now works properly
+- PresenceUpdate and MessageUpdate event now works properly
 
 ## [0.5.2] - 2020-05-26
 ### Added